requirements.txt
python-dotenv==1.0.0
openai==0.27.5
pyyaml
langchain==0.0.166
pyowm